Modified to gather the Product DisplayName and SharedDllRefCount, among other things.
Changed to gather the Product's DisplayName based on the component's productid and SharedDll RefCount based on the component's path.
Changed the componentsRoot registry path to be a const instead of passing it around, since it does not change.
Made some changes to better handle null values.
Changed to compile to an x86 executable since  the goal is to find info about x86 dlls.